This commit was manufactured by cvs2svn to create tag 'R2_0_5'.

+++ src/qli/lex.cpp
@@ -37,27 +37,33 @@
 #include "../qli/hsh_proto.h"
 #include "../qli/lex_proto.h"
 #include "../qli/proc_proto.h"
+#include "../jrd/gds_proto.h"
 #include "../jrd/utl_proto.h"
 #include "../jrd/gdsassert.h"
 #include "../common/utils_proto.h"
-#include "../common/classes/TempFile.h"
-
-using MsgFormat::SafeArg;
-
 
 #ifdef HAVE_UNISTD_H
 #include <unistd.h>
 #endif
 
-//#ifdef HAVE_CTYPES_H
-//#include <ctypes.h>
-//#endif
+#ifdef HAVE_CTYPES_H
+#include <ctypes.h>
+#endif
+
+#ifdef VMS
+#include <descrip.h>
+const SLONG LIB$_INPSTRTRU	= 0x15821c;
+#endif
 
 #if (defined WIN_NT)
 #include <io.h> // isatty
 #endif
 
+#ifdef SMALL_FILE_NAMES
+static const char* SCRATCH		= "fb_q";
+#else
 static const char* SCRATCH		= "fb_query_";
+#endif
 
 const char* FOPEN_INPUT_TYPE	= "r";
 
@@ -66,7 +72,7 @@ static int nextchar(const bool);
 static void next_line(const bool);
 static void retchar();
 static bool scan_number(SSHORT, TEXT **);
-static int skip_white();
+static int skip_white(void);
 
 static qli_lls* QLI_statements;
 static int QLI_position;
@@ -84,8 +90,8 @@ const char CHR_white	= char(16);
 const char CHR_eol		= char(32);
 
 const char CHR_IDENT	= CHR_ident;
-const char CHR_LETTER	= CHR_letter | CHR_ident;
-const char CHR_DIGIT	= CHR_digit | CHR_ident;
+const char CHR_LETTER	= CHR_letter + CHR_ident;
+const char CHR_DIGIT	= CHR_digit + CHR_ident;
 const char CHR_QUOTE	= CHR_quote;
 const char CHR_WHITE	= CHR_white;
 const char CHR_EOL		= CHR_eol;
@@ -93,9 +99,7 @@ const char CHR_EOL		= CHR_eol;
 static const char* const eol_string = "end of line";
 static const char* const eof_string = "*end_of_file*";
 
-// Do not reference the array directly; use the functions below.
-
-static const char classes_array[256] =
+static const char classes[256] =
 {
 	0, 0, 0, 0, 0, 0, 0, 0,
 	0, CHR_WHITE, CHR_EOL, 0, 0, 0, 0, 0,
@@ -122,19 +126,9 @@ static const char classes_array[256] =
 	CHR_LETTER, CHR_LETTER, CHR_LETTER, 0
 };
 
-inline char classes(int idx)
-{
-	return classes_array[(UCHAR) idx];
-}
-
-inline char classes(UCHAR idx)
-{
-	return classes_array[idx];
-}
 
 
-
-bool LEX_active_procedure()
+bool LEX_active_procedure(void)
 {
 /**************************************
  *
@@ -153,7 +147,7 @@ bool LEX_active_procedure()
 }
 
 
-void LEX_edit(SLONG start, SLONG stop)
+void LEX_edit( SLONG start, SLONG stop)
 {
 /**************************************
  *
@@ -166,17 +160,17 @@ void LEX_edit(SLONG start, SLONG stop)
  *	push the scratch file on the input stack.
  *
  **************************************/
-	const Firebird::PathName filename = TempFile::create(SCRATCH);
-	FILE* scratch = fopen(filename.c_str(), "w+b");
-	if (!scratch)
-		IBERROR(61);			// Msg 61 couldn't open scratch file
+	TEXT filename[MAXPATHLEN];
 
+	FILE* scratch = (FILE*) gds__temp_file(TRUE, SCRATCH, filename);
+	if (scratch == (FILE *) - 1)
+		IBERROR(61);			// Msg 61 couldn't open scratch file
 #ifdef WIN_NT
 	stop--;
 #endif
 
 	if (fseek(trace_file, start, 0)) {
-		fseek(trace_file, 0, 2);
+		fseek(trace_file, (SLONG) 0, 2);
 		IBERROR(59);			// Msg 59 fseek failed
 	}
 
@@ -189,16 +183,16 @@ void LEX_edit(SLONG start, SLONG stop)
 
 	fclose(scratch);
 
-	if (gds__edit(filename.c_str(), TRUE))
-		LEX_push_file(filename.c_str(), true);
+	if (gds__edit(filename, TRUE))
+		LEX_push_file(filename, true);
 
-	unlink(filename.c_str());
+	unlink(filename);
 
-	fseek(trace_file, 0, 2);
+	fseek(trace_file, (SLONG) 0, 2);
 }
 
 
-qli_tok* LEX_edit_string()
+qli_tok* LEX_edit_string(void)
 {
 /**************************************
  *
@@ -225,13 +219,10 @@ qli_tok* LEX_edit_string()
 		return NULL;
 	}
 
-	while (!(classes(c) & (CHR_white | CHR_eol)))
-	{
+	while (!(classes[c] & (CHR_white | CHR_eol))) {
 		*p++ = c;
-		if (classes(c) & CHR_quote)
-		{
-			for (;;)
-			{
+		if (classes[c] & CHR_quote)
+			for (;;) {
 				const SSHORT d = nextchar(false);
 				if (d == '\n') {
 					retchar();
@@ -241,7 +232,6 @@ qli_tok* LEX_edit_string()
 				if (d == c)
 					break;
 			}
-		}
 		c = nextchar(true);
 	}
 
@@ -264,7 +254,7 @@ qli_tok* LEX_edit_string()
 }
 
 
-qli_tok* LEX_filename()
+qli_tok* LEX_filename(void)
 {
 /**************************************
  *
@@ -297,7 +287,7 @@ qli_tok* LEX_filename()
 // notice if this looks like a quoted file name
 
 	SSHORT save = 0;
-	if (classes(c) & CHR_quote) {
+	if (classes[c] & CHR_quote) {
 		token->tok_type = tok_quoted;
 		save = c;
 	}
@@ -306,14 +296,12 @@ qli_tok* LEX_filename()
 
 	for (;;) {
 		c = nextchar(true);
-		char char_class = classes(c);
-		if (c == '"' && c != save)
-		{
+		char char_class = classes[c];
+		if (c == '"' && c != save) {
 			*p++ = c;
-			for (;;)
-			{
+			for (;;) {
 				c = nextchar(true);
-				char_class = classes(c);
+				char_class = classes[c];
 				if ((char_class & CHR_eol) || c == '"')
 					break;
 				*p++ = c;
@@ -351,7 +339,7 @@ qli_tok* LEX_filename()
 }
 
 
-void LEX_fini()
+void LEX_fini(void)
 {
 /**************************************
  *
@@ -364,14 +352,14 @@ void LEX_fini()
  *
  **************************************/
 
-	if (trace_file) {
+	if (trace_file && (trace_file != (FILE *) - 1)) {
 		fclose(trace_file);
 		unlink(trace_file_name);
 	}
 }
 
 
-void LEX_flush()
+void LEX_flush(void)
 {
 /**************************************
  *
@@ -397,25 +385,22 @@ void LEX_flush()
 // Look for a semi-colon
 
 	if (QLI_semi)
-	{
 		while (QLI_line && QLI_token->tok_keyword != KW_SEMI)
 			LEX_token();
-	}
 	else
-	{
 		while (QLI_line && QLI_token->tok_type != tok_eol)
 			LEX_token();
-	}
 }
 
 
+#if defined(UNIX) || defined(WIN_NT)
 bool LEX_get_line(const TEXT* prompt,
 				  TEXT * buffer,
 				  int size)
 {
 /**************************************
  *
- *	L E X _ g e t _ l i n e
+ *	L E X _ g e t _ l i n e  ( U N I X )
  *
  **************************************
  *
@@ -455,8 +440,8 @@ bool LEX_get_line(const TEXT* prompt,
 			}
 			if (QLI_abort)
 				continue;
-
-			break;
+			else
+				break;
 		}
 		if (--size > 0)
 			*p++ = c;
@@ -480,9 +465,74 @@ bool LEX_get_line(const TEXT* prompt,
 
 	return true;
 }
+#endif
 
 
-void LEX_init()
+#ifdef VMS
+bool LEX_get_line(const TEXT* prompt,
+				  TEXT * buffer,
+				  int size)
+{
+/**************************************
+ *
+ *	L E X _ g e t _ l i n e  ( V M S )
+ *
+ **************************************
+ *
+ * Functional description
+ *	Give a prompt and read a line.  If the line is terminated by
+ *	an EOL, return true.  If the buffer is exhausted and non-blanks
+ *	would be discarded, return an error.  If EOF is detected,
+ *	return false.  Regardless, a null terminated string is returned.
+ *
+ **************************************/
+	struct dsc$descriptor_s line_desc, prompt_desc;
+	SLONG status;
+	SSHORT length;
+
+// We're going to add a null to the end, so don't read too much
+
+	--size;
+
+	line_desc.dsc$b_class = DSC$K_CLASS_S;
+	line_desc.dsc$b_dtype = DSC$K_DTYPE_T;
+	line_desc.dsc$w_length = size;
+	line_desc.dsc$a_pointer = buffer;
+
+	if (prompt) {
+		prompt_desc.dsc$b_class = DSC$K_CLASS_S;
+		prompt_desc.dsc$b_dtype = DSC$K_DTYPE_T;
+		prompt_desc.dsc$w_length = strlen(prompt);
+		prompt_desc.dsc$a_pointer = const_cast<TEXT*>(prompt); // safe cast
+		status = lib$get_input(&line_desc, &prompt_desc, &length);
+	}
+	else
+		status = lib$get_input(&line_desc, NULL, &length);
+
+	SCHAR* p = buffer + length;
+
+	if (!(status & 1)) {
+		if (status != LIB$_INPSTRTRU)
+			return false;
+		buffer[0] = 0;
+		IBERROR(476);			// Msg 476 input line too long
+	}
+	else if (length < size)
+		*p++ = '\n';
+
+	*p = 0;
+
+	if (sw_verify) {
+		line_desc.dsc$w_length = length;
+		lib$put_output(&line_desc);
+	}
+
+	return true;
+}
+#endif
+
+
+void LEX_init(void)
 {
 /**************************************
  *
@@ -495,10 +545,8 @@ void LEX_init()
  *	scratch trace file to keep all input.
  *
  **************************************/
-	const Firebird::PathName filename = TempFile::create(SCRATCH);
-	strcpy(trace_file_name, filename.c_str());
-	trace_file = fopen(trace_file_name, "w+b");
-	if (!trace_file)
+	trace_file = (FILE*) gds__temp_file(TRUE, SCRATCH, trace_file_name);
+	if (trace_file == (FILE *) - 1)
 		IBERROR(61);			// Msg 61 couldn't open scratch file
 
 	QLI_token = (qli_tok*) ALLOCPV(type_tok, MAXSYMLEN);
@@ -515,7 +563,7 @@ void LEX_init()
 }
 
 
-void LEX_mark_statement()
+void LEX_mark_statement(void)
 {
 /**************************************
  *
@@ -530,7 +578,9 @@ void LEX_mark_statement()
  **************************************/
 	qli_line* temp;
 
-	for (temp = QLI_line; temp->line_next && QLI_statements; temp = temp->line_next)
+	for (temp = QLI_line;
+		 temp->line_next && QLI_statements;
+		 temp = temp->line_next)
 	{
 		if (temp->line_next->line_position == (IPTR) QLI_statements->lls_object)
 			return;
@@ -543,7 +593,7 @@ void LEX_mark_statement()
 }
 
 
-void LEX_pop_line()
+void LEX_pop_line(void)
 {
 /**************************************
  *
@@ -583,7 +633,8 @@ void LEX_procedure( DBB database, FB_API_HANDLE blob)
  **************************************/
 	qli_line* temp = (qli_line*) ALLOCPV(type_line, QLI_token->tok_length);
 	temp->line_source_blob = blob;
-	strncpy(temp->line_source_name, QLI_token->tok_string, QLI_token->tok_length);
+	strncpy(temp->line_source_name, QLI_token->tok_string,
+			QLI_token->tok_length);
 	temp->line_type = line_blob;
 	temp->line_database = database;
 	temp->line_size = sizeof(temp->line_data);
@@ -615,7 +666,7 @@ bool LEX_push_file(const TEXT* filename,
 		sprintf(buffer, "%s.com", filename);
 		if (!(file = fopen(buffer, FOPEN_INPUT_TYPE))) {
 			if (error_flag)
-				ERRQ_msg_put(67, filename);
+				ERRQ_msg_put(67, filename, NULL, NULL, NULL, NULL);
 				// Msg 67 can't open command file \"%s\"\n
 			return false;
 		}
@@ -675,7 +726,7 @@ void LEX_put_procedure(FB_API_HANDLE blob, SLONG start, SLONG stop)
 	ISC_STATUS_ARRAY status_vector;
 
 	if (fseek(trace_file, start, 0)) {
-		fseek(trace_file, 0, 2);
+		fseek(trace_file, (SLONG) 0, 2);
 		IBERROR(62);			// Msg 62 fseek failed
 	}
 
@@ -700,15 +751,16 @@ void LEX_put_procedure(FB_API_HANDLE blob, SLONG start, SLONG stop)
 			}
 		}
 		const SSHORT l = p - buffer;
-		if (l && isc_put_segment(status_vector, &blob, l, buffer))
-			ERRQ_bugcheck(58);	// Msg 58 isc_put_segment failed
+		if (l)
+			if (isc_put_segment(status_vector, &blob, l, buffer))
+				ERRQ_bugcheck(58);	// Msg 58 isc_put_segment failed
 	}
 
-	fseek(trace_file, 0, 2);
+	fseek(trace_file, (SLONG) 0, 2);
 }
 
 
-void LEX_real()
+void LEX_real(void)
 {
 /**************************************
  *
@@ -726,7 +778,7 @@ void LEX_real()
 }
 
 
-qli_lls* LEX_statement_list()
+qli_lls* LEX_statement_list(void)
 {
 /**************************************
  *
@@ -745,7 +797,7 @@ qli_lls* LEX_statement_list()
 }
 
 
-qli_tok* LEX_token()
+qli_tok* LEX_token(void)
 {
 /**************************************
  *
@@ -785,14 +837,15 @@ qli_tok* LEX_token()
 	}
 
 	*p++ = c;
-	QLI_token->tok_position = QLI_line->line_position + QLI_line->line_ptr - QLI_line->line_data - 1;
+	QLI_token->tok_position = QLI_line->line_position +
+		QLI_line->line_ptr - QLI_line->line_data - 1;
 
 // On end of file, generate furious but phone end of line tokens
 
-	char char_class = classes(c);
+	char char_class = classes[c];
 
 	if (char_class & CHR_letter) {
-		for (c = nextchar(true); classes(c) & CHR_ident; c = nextchar(true))
+		for (c = nextchar(true); classes[c] & CHR_ident; c = nextchar(true))
 			*p++ = c;
 		retchar();
 		token->tok_type = tok_ident;
@@ -810,7 +863,7 @@ qli_tok* LEX_token()
 			}
 			*p++ = next;
 			if ((p - token->tok_string) >= MAXSYMLEN)
-				ERRQ_msg_put(470, SafeArg() << MAXSYMLEN);	// Msg 470 literal too long
+				ERRQ_msg_put(470, (TEXT *) (IPTR) MAXSYMLEN, NULL, NULL, NULL, NULL);	// Msg 470 literal too long
 
 			// If there are 2 quotes in a row, interpret 2nd as a literal
 
@@ -904,8 +957,8 @@ static bool get_line(FILE * file,
 			}
 			if (QLI_abort)
 				continue;
-
-			break;
+			else
+				break;
 		}
 		if (--length > 0)
 			*p++ = c;
@@ -980,7 +1033,8 @@ static void next_line(const bool eof_ok)
 		if (QLI_line->line_type == line_blob) {
 			// If the current blob segment contains another line, use it
 
-			if ((p = QLI_line->line_ptr) != QLI_line->line_data && p[-1] == '\n' && *p)
+			if ((p = QLI_line->line_ptr) != QLI_line->line_data
+				&& p[-1] == '\n' && *p)
 			{
 				flag = true;
 			}
@@ -990,7 +1044,8 @@ static void next_line(const bool eof_ok)
 				p = QLI_line->line_data;
 				QLI_line->line_ptr = QLI_line->line_data;
 
-				flag = PRO_get_line(QLI_line->line_source_blob, p, QLI_line->line_size);
+				flag = PRO_get_line(QLI_line->line_source_blob, p,
+								 QLI_line->line_size);
 				if (flag && QLI_echo)
 					printf("%s", QLI_line->line_data);
 			}
@@ -1010,7 +1065,7 @@ static void next_line(const bool eof_ok)
 			}
 			if (flag) {
 				TEXT* q;
-				for (q = p; classes(*q) & CHR_white; q++);
+				for (q = p; classes[static_cast<UCHAR>(*q)] & CHR_white; q++);
 				if (*q == '@') {
 					TEXT filename[MAXPATHLEN];
 					for (p = q + 1, q = filename; *p && *p != '\n';)
@@ -1041,17 +1096,14 @@ static void next_line(const bool eof_ok)
 		// this is an unexpected end of file
 
 		if (QLI_line->line_type == line_blob)
-		{
-			ERRQ_print_error(64, QLI_line->line_source_name);
-			// Msg 64 unexpected end of procedure in procedure %s
-		}
+			ERRQ_print_error(64, QLI_line->line_source_name, NULL, NULL, NULL,
+							 NULL);
+		// Msg 64 unexpected end of procedure in procedure %s
 		else if (QLI_line->line_type == line_file)
-		{
-			ERRQ_print_error(65, QLI_line->line_source_name);
-			// Msg 65 unexpected end of file in file %s
-		}
-		else
-		{
+			ERRQ_print_error(65, QLI_line->line_source_name, NULL, NULL, NULL,
+							 NULL);
+		// Msg 65 unexpected end of file in file %s
+		else {
 			if (QLI_line->line_type == line_string)
 				LEX_pop_line();
 			IBERROR(66);		// Msg 66 unexpected eof
@@ -1066,12 +1118,9 @@ static void next_line(const bool eof_ok)
 // Dump output to the trace file
 
 	if (QLI_line->line_type == line_blob)
-	{
 		while (*p)
 			p++;
-	}
-	else
-	{
+	else {
 		while (*p)
 			putc(*p++, trace_file);
 		QLI_position += (TEXT *) p - QLI_line->line_data;
@@ -1121,24 +1170,24 @@ static bool scan_number(SSHORT c,
  **************************************/
 	bool dot = false;
 
+	TEXT* p = *ptr;
+
 // If this is a leading decimal point, check that the next
 //   character is really a digit, otherwise backout
 
 	if (c == '.') {
 		c = nextchar(true);
 		retchar();
-		if (!(classes(c) & CHR_digit))
+		if (!(classes[c] & CHR_digit))
 			return false;
 		dot = true;
 	}
-
-	TEXT* p = *ptr;
 
 // Gobble up digits up to a single decimal point
 
 	for (;;) {
 		c = nextchar(true);
-		if (classes(c) & CHR_digit)
+		if (classes[c] & CHR_digit)
 			*p++ = c;
 		else if (!dot && c == '.') {
 			*p++ = c;
@@ -1157,7 +1206,7 @@ static bool scan_number(SSHORT c,
 			*p++ = c;
 			c = nextchar(true);
 		}
-		while (classes(c) & CHR_digit) {
+		while (classes[c] & CHR_digit) {
 			*p++ = c;
 			c = nextchar(true);
 		}
@@ -1170,7 +1219,7 @@ static bool scan_number(SSHORT c,
 }
 
 
-static int skip_white()
+static int skip_white(void)
 {
 /**************************************
  *
@@ -1186,7 +1235,7 @@ static int skip_white()
 
 	while (true) {
 		c = nextchar(true);
-		const char char_class = classes(c);
+		const char char_class = classes[c];
 		if (char_class & CHR_white)
 			continue;
 		if (c == '/') {
